+++ test/Audit.IntegrationTest/ElasticsearchTests.cs
@@ -29,9 +29,10 @@ namespace Audit.IntegrationTest
             var ins = new List<Core.AuditEvent>();
             var repl = new List<Core.AuditEvent>();
             var ela = GetElasticsearchDataProvider(ins, repl);
-
+            var indexName = "auditevent_order";
+            
             var guids = new List<string>();
-            ela.IndexBuilder = ev => "auditevent2";
+            ela.IndexBuilder = ev => indexName;
             ela.IdBuilder = ev => { var g = Guid.NewGuid().ToString().Replace("-", "/"); guids.Add(g); return g; };
 
             Audit.Core.Configuration.Setup()
@@ -39,28 +40,89 @@ namespace Audit.IntegrationTest
                 .WithCreationPolicy(Core.EventCreationPolicy.InsertOnStartReplaceOnEnd)
                 .ResetActions();
 
-            var sb = "init";
+            var order = new Order()
+            {
+                Id = 1,
+                Status = "Created"
+            };
             
 
-            using (var scope = new AuditScopeFactory().Create("eventType", () => sb, new { MyCustomField = "value" }, null, null))
+            using (var scope = new AuditScopeFactory().Create("eventType", () => order, new { MyCustomField = "value" }, null, null))
             {
-                sb += "-end";
+                order.Status = "Updated";
             }
 
-            ela.Client.Indices.Refresh("auditevent2");
+            ela.Client.Indices.Refresh(indexName);
 
-            var evLoad = ela.GetEvent(new ElasticsearchAuditEventId() { Id = guids[0], Index = "auditevent2" });
+            var evLoad = ela.GetEvent(new ElasticsearchAuditEventId() { Id = guids[0], Index = indexName });
+            var orderOldValue = Configuration.JsonAdapter.Deserialize<Order>(repl[0].Target.Old.ToString());
+            var orderNewValue = Configuration.JsonAdapter.Deserialize<Order>(repl[0].Target.New.ToString());
+            var oldDictionary = evLoad.Target.Old as Dictionary<string, object>;
+            var newDictionary = evLoad.Target.New as Dictionary<string, object>;
 
             Assert.IsNotNull(evLoad);
+            Assert.IsNotNull(oldDictionary);
+            Assert.IsNotNull(newDictionary);
             Assert.AreEqual(1, guids.Count);
             Assert.AreEqual(1, ins.Count);
             Assert.AreEqual(1, repl.Count);
-            Assert.AreEqual("init", ins[0].Target.Old.ToString());
-            Assert.AreEqual("init", evLoad.Target.Old.ToString());
+            Assert.AreEqual("Created", orderOldValue.Status);
+            Assert.AreEqual("Created", oldDictionary["status"].ToString());
+            Assert.AreEqual("Updated", orderNewValue.Status);
+            Assert.AreEqual("Updated", newDictionary["status"].ToString());
+            Assert.AreEqual("value", evLoad.CustomFields["MyCustomField"]);
             Assert.AreEqual(null, ins[0].Target.New);
-            Assert.AreEqual("init", repl[0].Target.Old.ToString());
-            Assert.AreEqual("init-end", repl[0].Target.New.ToString());
-            Assert.AreEqual("init-end", evLoad.Target.New.ToString());
+        }
+
+        [Test]
+        [Category("Elasticsearch")]
+        public async Task Test_Elasticsearch_HappyPath_Async()
+        {
+            var ins = new List<Core.AuditEvent>();
+            var repl = new List<Core.AuditEvent>();
+            var ela = GetElasticsearchDataProvider(ins, repl);
+            var indexName = "auditevent_order";
+
+            var guids = new List<string>();
+            ela.IndexBuilder = ev => indexName;
+            ela.IdBuilder = ev => { var g = Guid.NewGuid().ToString().Replace("-", "/"); guids.Add(g); return g; };
+
+            Audit.Core.Configuration.Setup()
+                .UseCustomProvider(ela)
+                .WithCreationPolicy(Core.EventCreationPolicy.InsertOnStartReplaceOnEnd)
+                .ResetActions();
+
+            var order = new Order()
+            {
+                Id = 1,
+                Status = "Created"
+            };
+
+            using (var scope = await new AuditScopeFactory().CreateAsync("eventType", () => order, new { MyCustomField = "value" }, null, null))
+            {
+                order.Status = "Updated";
+            }
+
+            await ela.Client.Indices.RefreshAsync(indexName);
+
+            var evLoad = await ela.GetEventAsync(new ElasticsearchAuditEventId() { Id = guids[0], Index = indexName });
+            var orderOldValue = Configuration.JsonAdapter.Deserialize<Order>(repl[0].Target.Old.ToString());
+            var orderNewValue = Configuration.JsonAdapter.Deserialize<Order>(repl[0].Target.New.ToString());
+            var oldDictionary = evLoad.Target.Old as Dictionary<string, object>;
+            var newDictionary = evLoad.Target.New as Dictionary<string, object>;
+
+            Assert.IsNotNull(evLoad);
+            Assert.IsNotNull(oldDictionary);
+            Assert.IsNotNull(newDictionary);
+            Assert.AreEqual(1, guids.Count);
+            Assert.AreEqual(1, ins.Count);
+            Assert.AreEqual(1, repl.Count);
+            Assert.AreEqual("Created", orderOldValue.Status);
+            Assert.AreEqual("Created", oldDictionary["status"].ToString());
+            Assert.AreEqual("Updated", orderNewValue.Status);
+            Assert.AreEqual("Updated", newDictionary["status"].ToString());
+            Assert.AreEqual("value", evLoad.CustomFields["MyCustomField"]);
+            Assert.AreEqual(null, ins[0].Target.New);
         }
 
         [Test]
@@ -108,47 +170,6 @@ namespace Audit.IntegrationTest
             Assert.AreEqual("init-end", repl[0].Target.New.ToString());
             Assert.AreEqual("init-end", evResult.Target.New.ToString());
             Assert.AreEqual("value", evResult.CustomFields["MyCustomField"]?.ToString());
-        }
-
-        [Test]
-        [Category("Elasticsearch")]
-        public async Task Test_Elasticsearch_HappyPath_Async()
-        {
-            var ins = new List<Core.AuditEvent>();
-            var repl = new List<Core.AuditEvent>();
-            var ela = GetElasticsearchDataProvider(ins, repl);
-
-            var guids = new List<string>();
-            ela.IndexBuilder = ev => "auditevent2";
-            ela.IdBuilder = ev => { var g = Guid.NewGuid().ToString().Replace("-", "/"); guids.Add(g); return g; };
-
-            Audit.Core.Configuration.Setup()
-                .UseCustomProvider(ela)
-                .WithCreationPolicy(Core.EventCreationPolicy.InsertOnStartReplaceOnEnd)
-                .ResetActions();
-
-            var sb = "init";
-
-
-            using (var scope = await new AuditScopeFactory().CreateAsync("eventType", () => sb, new { MyCustomField = "value" }, null, null))
-            {
-                sb += "-end";
-            }
-
-            await ela.Client.Indices.RefreshAsync("auditevent2");
-
-            var evLoad = await ela.GetEventAsync(new ElasticsearchAuditEventId() { Id = guids[0], Index = "auditevent2" });
-
-            Assert.IsNotNull(evLoad);
-            Assert.AreEqual(1, guids.Count);
-            Assert.AreEqual(1, ins.Count);
-            Assert.AreEqual(1, repl.Count);
-            Assert.AreEqual("init", ins[0].Target.Old.ToString());
-            Assert.AreEqual("init", evLoad.Target.Old.ToString());
-            Assert.AreEqual(null, ins[0].Target.New);
-            Assert.AreEqual("init", repl[0].Target.Old.ToString());
-            Assert.AreEqual("init-end", repl[0].Target.New.ToString());
-            Assert.AreEqual("init-end", evLoad.Target.New.ToString());
         }
 
         [Test]