+++ test/Audit.RavenDB.UnitTest/RavenDbDataProviderTests.cs
@@ -104,8 +104,8 @@ namespace Audit.RavenDB.UnitTest
             Assert.AreEqual("sub-field value", ((dynamic)ev.CustomFields["doc"])?.subField?.ToString());
             Assert.AreEqual("test", ((dynamic)ev.Environment.CustomFields["extra"])?.prop.ToString());
 #endif
-            Assert.AreEqual(20, (ev.Target.Old as JToken)?["LowTemp"]?.Value<int>());
-            Assert.AreEqual(21, (ev.Target.New as JToken)?["LowTemp"]?.Value<int>());
+            Assert.AreEqual(20, (ev.Target.Old as DailyTemperature)?.LowTemp);
+            Assert.AreEqual(21, (ev.Target.New as DailyTemperature)?.LowTemp);
         }
 
         [Test]
@@ -150,8 +150,8 @@ namespace Audit.RavenDB.UnitTest
             Assert.AreEqual("sub-field value", ((dynamic)ev.CustomFields["doc"])?.subField?.ToString());
             Assert.AreEqual("test", ((dynamic)ev.Environment.CustomFields["extra"])?.prop.ToString());
 #endif
-            Assert.AreEqual(20, (ev.Target.Old as JToken)?["LowTemp"]?.Value<int>());
-            Assert.AreEqual(21, (ev.Target.New as JToken)?["LowTemp"]?.Value<int>());
+            Assert.AreEqual(20, (ev.Target.Old as DailyTemperature)?.LowTemp);
+            Assert.AreEqual(21, (ev.Target.New as DailyTemperature)?.LowTemp);
         }
 
 #if IS_TEXT_JSON
@@ -192,8 +192,8 @@ namespace Audit.RavenDB.UnitTest
             Assert.IsNotNull(ev);
             Assert.AreEqual("field value", ev.CustomFields["field"].ToString());
             Assert.AreEqual("sub-field value", (ev.CustomFields["doc"] as JToken)?["subField"]?.ToString());
-            Assert.AreEqual(20, (ev.Target.Old as JToken)?["LowTemp"]?.Value<int>());
-            Assert.AreEqual(21, (ev.Target.New as JToken)?["LowTemp"]?.Value<int>());
+            Assert.AreEqual(20, (ev.Target.Old as DailyTemperature)?.LowTemp);
+            Assert.AreEqual(21, (ev.Target.New as DailyTemperature)?.LowTemp);
         }
 #endif
 
@@ -244,8 +244,8 @@ namespace Audit.RavenDB.UnitTest
             Assert.AreEqual("sub-field value", ((dynamic)ev.CustomFields["doc"])?.subField?.ToString());
             Assert.AreEqual("test", ((dynamic)ev.Environment.CustomFields["extra"])?.prop.ToString());
 #endif
-            Assert.AreEqual(20, (ev.Target.Old as JToken)?["LowTemp"]?.Value<int>());
-            Assert.AreEqual(21, (ev.Target.New as JToken)?["LowTemp"]?.Value<int>());
+            Assert.AreEqual(20, (ev.Target.Old as DailyTemperature)?.LowTemp);
+            Assert.AreEqual(21, (ev.Target.New as DailyTemperature)?.LowTemp);
         }
 
         [Test]
@@ -303,10 +303,10 @@ namespace Audit.RavenDB.UnitTest
             Assert.AreEqual("sub-field value", ((dynamic)ev2.CustomFields["doc"])?.subField?.ToString());
             Assert.AreEqual("test", ((dynamic)ev2.Environment.CustomFields["extra"])?.prop.ToString());
 #endif
-            Assert.AreEqual(20, (ev1.Target.Old as JToken)?["LowTemp"]?.Value<int>());
+            Assert.AreEqual(20, (ev1.Target.Old as DailyTemperature)?.LowTemp);
             Assert.IsNull(ev1.Target.New);
-            Assert.AreEqual(20, (ev2.Target.Old as JToken)?["LowTemp"]?.Value<int>());
-            Assert.AreEqual(21, (ev2.Target.New as JToken)?["LowTemp"]?.Value<int>());
+            Assert.AreEqual(20, (ev2.Target.Old as DailyTemperature)?.LowTemp);
+            Assert.AreEqual(21, (ev2.Target.New as DailyTemperature)?.LowTemp);
         }